Respect silent switch in and out of app.
// FREEBIE
