+++ Signal/src/view controllers/CallViewController.swift
@@ -6,117 +6,6 @@ import Foundation
 import WebRTC
 import PromiseKit
 
-// TODO move this somewhere else.
-@objc class CallAudioService: NSObject {
-    private let TAG = "[CallAudioService]"
-    private var vibrateTimer: Timer?
-    private let audioManager = AppAudioManager.sharedInstance()
-    private let soundPlayer = JSQSystemSoundPlayer.shared()!
-
-    enum SoundFilenames: String {
-        case incomingRing = "r"
-    }
-
-    // Mark: Vibration config
-    private let vibrateRepeatDuration = 1.6
-
-    // Our ring buzz is a pair of vibrations.
-    // `pulseDuration` is the small pause between the two vibrations in the pair.
-    private let pulseDuration = 0.2
-
-    public var isSpeakerphoneEnabled = false {
-        didSet {
-            handleUpdatedSpeakerphone()
-        }
-    }
-
-    public func handleState(_ state: CallState) {
-        switch state {
-        case .idle: handleIdle()
-        case .dialing: handleDialing()
-        case .answering: handleAnswering()
-        case .remoteRinging: handleRemoteRinging()
-        case .localRinging: handleLocalRinging()
-        case .connected: handleConnected()
-        case .localFailure: handleLocalFailure()
-        case .localHangup: handleLocalHangup()
-        case .remoteHangup: handleRemoteHangup()
-        case .remoteBusy: handleBusy()
-        }
-    }
-
-    private func handleIdle() {
-        Logger.debug("\(TAG) \(#function)")
-    }
-
-    private func handleDialing() {
-        Logger.debug("\(TAG) \(#function)")
-    }
-
-    private func handleAnswering() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleRemoteRinging() {
-        Logger.debug("\(TAG) \(#function)")
-    }
-
-    private func handleLocalRinging() {
-        Logger.debug("\(TAG) \(#function)")
-
-        vibrateTimer = Timer.scheduledTimer(timeInterval: vibrateRepeatDuration, target: self, selector: #selector(ringVibration), userInfo: nil, repeats: true)
-        soundPlayer.playSound(withFilename: SoundFilenames.incomingRing.rawValue, fileExtension: kJSQSystemSoundTypeCAF)
-    }
-
-    private func handleConnected() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleLocalFailure() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleLocalHangup() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleRemoteHangup() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleBusy() {
-        Logger.debug("\(TAG) \(#function)")
-        stopRinging()
-    }
-
-    private func handleUpdatedSpeakerphone() {
-        audioManager.toggleSpeakerPhone(isEnabled: isSpeakerphoneEnabled)
-    }
-
-    // MARK: Helpers
-
-    private func stopRinging() {
-        vibrateTimer?.invalidate()
-        vibrateTimer = nil
-
-        soundPlayer.stopSound(withFilename: SoundFilenames.incomingRing.rawValue)
-    }
-
-    public func ringVibration() {
-        // Since a call notification is more urgent than a message notifaction, we
-        // vibrate twice, like a pulse, to differentiate from a normal notification vibration.
-        soundPlayer.playVibrateSound()
-        DispatchQueue.default.asyncAfter(deadline: DispatchTime.now() + pulseDuration) {
-            self.soundPlayer.playVibrateSound()
-        }
-    }
-}
-
 // TODO: Add category so that button handlers can be defined where button is created.
 // TODO: Add logic to button handlers.
 // TODO: Ensure buttons enabled & disabled as necessary.
